+++ pkg/k8sutil/apply_yaml.go
@@ -172,9 +172,6 @@ func (y *YAMLApplier) applyAction(obj *unstructured.Unstructured) error {
 		if err != nil {
 			return err
 		}
-		if fieldObj == nil {
-			continue
-		}
 		cf.typeOf = reflect.TypeOf(fieldObj).String()
 		clientFields = append(clientFields, cf)
 	}
