+++ platform-operator/experimental/controllers/platformctrl/common/semver_utils.go
@@ -1,6 +1,6 @@
 // Copyright (c) 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-package semver
+package common
 
 import "github.com/Masterminds/semver/v3"
